requirements.txt
numpy==1.14.3
tensorflow_gpu==1.8.0

